+++ index.js
@@ -32,17 +32,21 @@ export const ribbon = {
 };
 
 export const slider = {
-  wrapper: 'f-slider relative w-full h-44 py-2',
-  track: 'track absolute bg-blue-200 h-4 top-20 rounded-4 w-full',
-  trackDisabled: 'pointer-events-none bg-gray-200',
+  wrapper: 'touch-pan-y relative w-full h-44 py-2',
+  track:
+    'absolute i-bg-$color-slider-track-background h-4 top-20 rounded-4 w-full ',
+  trackDisabled:
+    'pointer-events-none i-bg-$color-slider-track-background-disabled',
   activeTrack:
-    'larger-active-track absolute bg-blue-600 h-4 top-20 rounded-4',
-  activeTrackDisabled: 'bg-bluegray-300 pointer-events-none',
+    'absolute i-bg-$color-slider-track-background-active h-6 top-[19px] rounded-4',
+  activeTrackDisabled:
+    'i-bg-$color-slider-track-background-disabled pointer-events-none',
   thumb:
-    'f-thumb absolute w-24 h-24 bottom-10 rounded-4 outline-none',
-  thumbEnabled: 'shadow bg-blue-600 cursor-pointer',
+    'absolute i-shadow-$shadow-slider transition-shadow w-24 h-24 bottom-10 rounded-4 hover:i-bg-$color-slider-handle-background-hover active:i-bg-$color-slider-handle-background-active active:slider-handle-shadow-active hover:slider-handle-shadow-hover',
+  thumbEnabled:
+    'i-bg-$color-slider-handle-background cursor-pointer',
   thumbDisabled:
-    'bg-bluegray-300 cursor-disabled pointer-events-none',
+    'i-bg-$color-slider-handle-background-disabled cursor-disabled pointer-events-none',
 };
 
 export const modal = {
@@ -369,7 +373,8 @@ export const toggle = {
   checkboxDisabled: 'before:i-bg-$color-checkbox-background-disabled before:i-border-$color-checkbox-border-disabled peer-checked:before:i-border-$color-checkbox-border-active-disabled peer-checked:before:i-bg-$color-checkbox-background-active-disabled',
   labelCheckboxBackground: 'i-bg-$color-checkbox-background',
   labelCheckboxBorder: 'i-border-$color-checkbox-border',
-  radio: 'before:rounded-full peer-checked:before:border-[6] peer-checked:peer-hover:before:i-border-$color-radio-border-active-hover peer-hover:before:i-border-$color-radio-border-hover peer-hover:before:i-bg-$color-radio-background-hover peer-checked:before:i-border-$color-radio-border-active',
+  radio: 'before:rounded-full peer-checked:before:border-[6] peer-checked:peer-hover:before:i-border-$color-radio-border-active-hover peer-hover:before:i-border-$color-radio-border-hover peer-hover:before:i-bg-$color-radio-background-hover',
+  radioChecked: 'peer-checked:before:i-border-$color-radio-border-active',
   radioInvalid: 'before:i-bg-$color-radio-negative-background peer-hover:before:i-bg-$color-radio-negative-background-hover before:i-border-$color-radio-negative-border peer-hover:before:i-border-$color-radio-negative-border-hover peer-checked:before:i-border-$color-radio-negative-border-active peer-checked:peer-hover:before:i-border-$color-radio-negative-border-active-hover ',
   radioDisabled: 'before:i-bg-$color-radio-background-disabled before:i-border-$color-radio-border-disabled peer-checked:before:i-border-$color-radio-border-active-disabled',
   labelRadioBackground: 'i-bg-$color-radio-background',
@@ -383,3 +388,8 @@ export const toggle = {
   icon: 'peer-checked:before:bg-center before:bg-[var(--x-form-check-mark)]',
   a11y: 'sr-only',
 };
+
+export const clickable = {
+  clickable: 'absolute inset-0 h-full w-full appearance-none cursor-pointer', 
+  label: `px-12 py-8 ${label.label} cursor-pointer`, 
+}